+++ src/EntityGraphQL/Compiler/GqlNodes/GraphQLDirective.cs
@@ -22,7 +22,7 @@ public class GraphQLDirective
     public Expression? Process(ISchemaProvider schema, Expression fieldExpression, IReadOnlyDictionary<string, object> args, ParameterExpression? docParam, object? docVariables)
     {
         var validationErrors = new List<string>();
-        var arguments = ArgumentUtil.BuildArgumentsObject(schema, name, null, inlineArgValues.MergeNew(args), processor.GetArguments(schema), processor.GetArgumentsType(), docParam, docVariables, validationErrors);
+        var arguments = ArgumentUtil.BuildArgumentsObject(schema, name, null, inlineArgValues.MergeNew(args), processor.GetArguments(schema).Values, processor.GetArgumentsType(), docParam, docVariables, validationErrors);
 
         if (validationErrors.Count > 0)
         {
@@ -35,7 +35,7 @@ public class GraphQLDirective
     public BaseGraphQLField? ProcessField(ISchemaProvider schema, BaseGraphQLField field, IReadOnlyDictionary<string, object> args, ParameterExpression? docParam, object? docVariables)
     {
         var validationErrors = new List<string>();
-        var arguments = ArgumentUtil.BuildArgumentsObject(schema, name, null, inlineArgValues.MergeNew(args), processor.GetArguments(schema), processor.GetArgumentsType(), docParam, docVariables, validationErrors);
+        var arguments = ArgumentUtil.BuildArgumentsObject(schema, name, null, inlineArgValues.MergeNew(args), processor.GetArguments(schema).Values, processor.GetArgumentsType(), docParam, docVariables, validationErrors);
 
         if (validationErrors.Count > 0)
         {