+++ src/EntityGraphQL/Schema/SchemaIntrospection.cs
@@ -348,7 +348,7 @@
             {
                 Description = directive.Description,
                 Locations = directive.On.Select(i => Enum.GetName(typeof(ExecutableDirectiveLocation), i))!,
-                Args = directive.GetArguments(schema).Select(arg => new InputValue(arg.Name, BuildType(schema, arg.Type, arg.Type.TypeDotnet, true))
+                Args = directive.GetArguments(schema).Values.Select(arg => new InputValue(arg.Name, BuildType(schema, arg.Type, arg.Type.TypeDotnet, true))
                 {
                     Description = arg.Description,
                     DefaultValue = null,