+++ src/tests/EntityGraphQL.Tests/QueryTests/InheritanceAdvancedTest.cs
@@ -1,10 +1,6 @@
 ﻿using EntityGraphQL.Schema;
-using System;
 using System.Collections.Generic;
-using System.Linq;
-using System.Text;
 using System.Text.Json;
-using System.Threading.Tasks;
 using Xunit;
 
 namespace EntityGraphQL.Tests